Module Tjr_plist_examples.Marshal_factory

type ('a, 'buf) marshaller = {

max_elt_sz : int;

m_elt : 'a -> ('buf * int) -> 'buf * int;

u_elt : 'buf -> int -> 'a * int;

}

Generic marshaller type, assuming max_elt_sz is known (alternative is to marshal and check for end of buffer explicitly)

module Make_marshaller : functor (X : sig ... end) -> sig ... end

Functor to construct a marshaller form a type deriving bin_io

module Internal : sig ... end

type arg =

| A1_int_option__ba_buf

type res =

| R1 of (int option, Internal.buf) marshaller

val make : arg -> res

val int_option_marshaller : (int option, Internal.buf) marshaller

Convenience